backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0
groq>=0.12.0
